+++ .storybook/preview.js
@@ -1,7 +1,7 @@
-import { GlobalStyles } from "./../src/styles/globalStyles";
+import { GlobalStyles } from './../src/styles/globalStyles';
 
 export const parameters = {
-  actions: { argTypesRegex: "^on[A-Z].*" },
+  actions: { argTypesRegex: '^on[A-Z].*' },
 };
 
 export const decorators = [