+++ src/components/Button/index.tsx
@@ -2,12 +2,10 @@ import React from 'react';
 import { IButtonProps } from './interfaces';
 import Container from './styles';
 
-const Button: React.FC<IButtonProps> = ({
-    backgroundColor,
-    color,
-    customStyles,
-    children,
-    ...props
-}) => <Container backgroundColor={backgroundColor} color={color} customStyles={customStyles} {...props}>{children}</Container>;
+const Button: React.FC<IButtonProps> = ({ backgroundColor, color, customStyles, children, ...props }) => (
+  <Container backgroundColor={backgroundColor} color={color} customStyles={customStyles} {...props}>
+    {children}
+  </Container>
+);
 
 export default Button;