+++ src/components/CustomButton/styles.ts
@@ -1,8 +1,8 @@
-import { HTMLProps } from "react";
-import styled, { css } from "styled-components";
+import { HTMLProps } from 'react';
+import styled, { css } from 'styled-components';
 
 export type ButtonProps = {
-  variant: "filled" | "outline" | "link" | "text";
+  variant: 'filled' | 'outline' | 'link' | 'text';
   small: boolean;
 } & HTMLProps<HTMLButtonElement>;
 
@@ -109,5 +109,5 @@ export const Button = styled.button<ButtonProps>`
   &:hover:not(:disabled) {
     background-color: #02ce87;
   }
-  ${({ variant = "filled" }) => buttonModifiers[variant]}
+  ${({ variant = 'filled' }) => buttonModifiers[variant]}
 `;