+++ src/components/Icons/IconList/index.stories.tsx
@@ -1,16 +1,16 @@
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
-import { Story, Meta } from "@storybook/react/types-6-0";
+import { Story, Meta } from '@storybook/react/types-6-0';
 
-import IconList from ".";
+import IconList from '.';
 
 export default {
-  title: "Icons",
+  title: 'Icons',
   component: IconList,
   argTypes: {
-    fill: { control: "color" },
+    fill: { control: 'color' },
   },
   args: {
-    fill: "#000000",
+    fill: '#000000',
   },
 } as Meta;