+++ src/components/Radio/index.stories.tsx
@@ -1,11 +1,11 @@
-import React from "react";
+import React from 'react';
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
-import { Story, Meta } from "@storybook/react/types-6-0";
+import { Story, Meta } from '@storybook/react/types-6-0';
 
-import Radio, { RadioProps } from "./";
+import Radio, { RadioProps } from './';
 
 export default {
-  title: "Radio",
+  title: 'Radio',
   component: Radio,
   argTypes: {},
 } as Meta;