+++ src/components/Select/Content/ItemSingle/index.tsx
@@ -2,18 +2,8 @@ import React from 'react';
 import { Container } from './styles';
 import { IItemSingle } from '../../interfaces';
 
-const ItemSingle: React.FC<IItemSingle> = ({
-  content,
-  handleSelected,
-  contentRender,
-  ...props
-}) => {
-
-  return (
-    <Container onClick={(e: any) => handleSelected(e, content)}>
-      {contentRender}
-    </Container>
-  )
-}
+const ItemSingle: React.FC<IItemSingle> = ({ content, handleSelected, contentRender, ...props }) => {
+  return <Container onClick={(e: any) => handleSelected(e, content)}>{contentRender}</Container>;
+};
 
 export default ItemSingle;