+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2021-05-21
+
+- Add eslint with CRA rules
+- Add prettier
+- Format code following new rules
+
 ## [1.1.4] - 2021-05-21
 
 - Add LICENSE
@@ -34,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.43] - 2021-05-18
 
 - Removed dangerouslySetInnerHTML from Loader
+
 ## [0.1.42] - 2021-05-12
 
 - Added Single Select
@@ -42,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.41] - 2021-05-12
 
 - Added Modal component
+
 ## [0.1.40] - 2021-05-11
 
 - Added Radio component